requirements.txt
PyQt5==5.15.9
telethon==1.36.0
pyrogram==2.0.106
pyperclip==1.9.0
pyinstaller==6.3.0
python-telegram-bot==20.8
pytz==2024.1
python-dotenv==1.0.0
